Separate native and web classes
