+++ android/shopify_buy_plugin/src/main/java/com/shopify/unity/buy/androidpay/view/widget/Updatable.java
@@ -1,4 +1,4 @@
-package com.shopify.unity.buy.view.widget;
+package com.shopify.unity.buy.androidpay.view.widget;
 
 /**
  * Defines that a view consumes an object that is used to update the view state.
