+++ android/shopify_buy_plugin/src/main/java/com/shopify/unity/buy/models/JsonSerializable.java
@@ -1,4 +1,4 @@
-package com.shopify.unity.buy;
+package com.shopify.unity.buy.models;
 
 import org.json.JSONException;
