+++ android/shopify_buy_plugin/src/main/java/com/shopify/unity/buy/web/WebIntentListener.java
@@ -1,4 +1,4 @@
-package com.shopify.unity.buy;
+package com.shopify.unity.buy.web;
 
 public interface WebIntentListener {
     void onWebIntentClose();
