+++ android/shopify_buy_plugin/src/main/res/layout/view_confirmation.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.shopify.unity.buy.view.widget.ConfirmationView
+<com.shopify.unity.buy.androidpay.view.widget.ConfirmationView
     android:id="@+id/confirmation"
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
@@ -68,4 +68,4 @@
             android:text="@string/confirmation_confirm" />
     </FrameLayout>
 
-</com.shopify.unity.buy.view.widget.ConfirmationView>
+</com.shopify.unity.buy.androidpay.view.widget.ConfirmationView>
