+++ android/shopify_buy_plugin/src/main/res/layout/view_confirmation_shipping_rates.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.shopify.unity.buy.view.widget.ShippingRatesView
+<com.shopify.unity.buy.androidpay.view.widget.ShippingRatesView
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
@@ -57,4 +57,4 @@
         app:layout_constraintTop_toBottomOf="@+id/change"
         tools:text="$59.60" />
 
-</com.shopify.unity.buy.view.widget.ShippingRatesView>
+</com.shopify.unity.buy.androidpay.view.widget.ShippingRatesView>
