+++ android/shopify_buy_plugin/src/main/res/layout/view_confirmation_total_summary.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.shopify.unity.buy.view.widget.TotalSummaryView
+<com.shopify.unity.buy.androidpay.view.widget.TotalSummaryView
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
@@ -111,4 +111,4 @@
         app:layout_constraintTop_toBottomOf="@+id/view"
         tools:text="Total: $20.00" />
 
-</com.shopify.unity.buy.view.widget.TotalSummaryView>
+</com.shopify.unity.buy.androidpay.view.widget.TotalSummaryView>
