+++ scripts/generator/graphql_generator/csharp/SDK/Android/AndroidWebCheckout.cs.erb
@@ -25,7 +25,7 @@ namespace <%= namespace %>.SDK.Android {
         }
 
         private const string UnityPlayerClassName = "com.unity3d.player.UnityPlayer";
-        private const string WebCheckoutSessionClassName = "com.shopify.unity.buy.WebCheckoutSession";
+        private const string WebCheckoutSessionClassName = "com.shopify.unity.buy.web.WebCheckoutSession";
 
         public AndroidWebCheckout(Cart cart, ShopifyClient client) {
             _cart = cart;
